kafka/server: Reduce header dependencies
