+++ src/v/kafka/protocol/tests/field_parser_test.cc
@@ -15,6 +15,7 @@
 #include "kafka/protocol/wire.h"
 #include "kafka/types.h"
 #include "random/generators.h"
+#include "utils/base64.h"
 
 #include <seastar/core/thread.hh>
 #include <seastar/testing/thread_test_case.hh>
