+++ src/v/kafka/server/fetch_session.h
@@ -14,7 +14,6 @@
 #include "model/fundamental.h"
 #include "model/ktp.h"
 #include "model/timeout_clock.h"
-#include "model/timestamp.h"
 
 #include <absl/container/flat_hash_map.h>
 #include <boost/iterator/iterator_adaptor.hpp>
