+++ src/v/kafka/server/group_recovery_consumer.cc
@@ -13,6 +13,7 @@
 
 #include "kafka/protocol/wire.h"
 #include "kafka/server/group_metadata.h"
+#include "kafka/server/logger.h"
 
 #include <exception>
