+++ src/v/kafka/server/group_router.cc
@@ -10,6 +10,8 @@
  */
 #include "kafka/server/group_router.h"
 
+#include "kafka/server/logger.h"
+
 #include <seastar/core/coroutine.hh>
 #include <seastar/core/when_all.hh>
 #include <seastar/core/with_scheduling_group.hh>
