+++ src/v/kafka/server/member.h
@@ -23,10 +23,8 @@
 
 #include <absl/container/flat_hash_set.h>
 
-#include <algorithm>
 #include <chrono>
 #include <iosfwd>
-#include <iterator>
 #include <memory>
 #include <optional>
 #include <utility>
