+++ src/v/kafka/server/request_context.h
@@ -13,14 +13,12 @@
 #include "base/seastarx.h"
 #include "base/vlog.h"
 #include "bytes/iobuf.h"
-#include "cluster/security_frontend.h"
 #include "kafka/protocol/fetch.h"
 #include "kafka/protocol/fwd.h"
 #include "kafka/protocol/types.h"
 #include "kafka/protocol/wire.h"
 #include "kafka/server/connection_context.h"
 #include "kafka/server/fetch_session_cache.h"
-#include "kafka/server/handlers/fetch/replica_selector.h"
 #include "kafka/server/handlers/handler_interface.h"
 #include "kafka/server/logger.h"
 #include "kafka/server/response.h"
@@ -30,7 +28,6 @@
 #include "model/namespace.h"
 #include "pandaproxy/schema_registry/fwd.h"
 #include "security/acl.h"
-#include "security/audit/schemas/application_activity.h"
 #include "security/audit/schemas/iam.h"
 #include "security/audit/schemas/types.h"
 #include "security/audit/schemas/utils.h"
