+++ src/v/kafka/server/rm_group_frontend.cc
@@ -25,6 +25,7 @@
 #include "kafka/server/coordinator_ntp_mapper.h"
 #include "kafka/server/group.h"
 #include "kafka/server/group_router.h"
+#include "kafka/server/logger.h"
 #include "rpc/connection_cache.h"
 
 #include <seastar/core/coroutine.hh>
