+++ src/v/kafka/server/token_bucket_rate_tracker.h
@@ -17,8 +17,6 @@
 #include <seastar/util/bool_class.hh>
 
 #include <chrono>
-#include <numeric>
-#include <vector>
 
 namespace kafka {
